+++ core-web/libs/sdk/analytics/package.json
@@ -20,18 +20,20 @@
     },
     "homepage": "https://github.com/dotCMS/core/tree/main/core-web/libs/sdk/analytics/README.md",
     "dependencies": {
-        "analytics": "^0.8.14",
-        "@analytics/storage-utils": "^0.4.2",
         "@analytics/activity-utils": "^0.1.16",
-        "@analytics/listener-utils": "^0.4.0"
+        "@analytics/listener-utils": "^0.4.0",
+        "@analytics/localstorage-utils": "^0.1.10",
+        "@analytics/session-storage-utils": "^0.0.7",
+        "@analytics/storage-utils": "^0.4.2",
+        "analytics": "^0.8.14"
     },
     "peerDependencies": {
         "react": "19.1.0"
     },
     "devDependencies": {
-        "vite": "~5.0.0",
         "@testing-library/jest-dom": "^6.1.6",
-        "@testing-library/react": "^14.0.0"
+        "@testing-library/react": "^14.0.0",
+        "vite": "~5.0.0"
     },
     "main": "./index.cjs.js",
     "module": "./index.esm.js",
